requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
requests==2.31.0
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Werkzeug==3.0.1
email-validator==2.1.0.post1
python-dotenv==1.0.0
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
alembic==1.13.1